+++ lib/vault/request.rb
@@ -15,5 +15,27 @@ module Vault
     def inspect
       "#<#{self.class.name}:0x#{"%x" % (self.object_id << 1)}>"
     end
+
+    private
+
+    # Removes the given header fields from options and returns the result. This
+    # modifies the given options in place.
+    #
+    # @param [Hash] options
+    #
+    # @return [Hash]
+    def extract_headers!(options = {})
+      extract = {
+        wrap_ttl: Vault::Client::WRAP_TTL_HEADER,
+      }
+
+      {}.tap do |h|
+        extract.each do |k,v|
+          if options[k]
+            h[v] = options.delete(k)
+          end
+        end
+      end
+    end
   end
 end