+++ spec/integration/api/logical_spec.rb
@@ -4,7 +4,7 @@ module Vault
   describe Logical do
     subject { vault_test_client.logical }
 
-    describe "#list" do
+    describe "#list", vault: ">= 0.5" do
       it "returns the empty array when no items exist" do
         expect(subject.list("secret/that/never/existed")).to eq([])
       end
@@ -83,6 +83,50 @@ module Vault
           subject.delete("secret/delete")
           subject.delete("secret/delete")
         }.to_not raise_error
+      end
+    end
+
+    describe "#unwrap", vault: ">= 0.6" do
+      it "returns the wrapped secret when it exists" do
+        wrapped = vault_test_client.auth_token.create(wrap_ttl: "5s")
+        unwrapped = subject.unwrap(wrapped.wrap_info.token)
+
+        expect(unwrapped.auth).to be
+        expect(unwrapped.auth.client_token).to be
+
+        vault_test_client.with_token(unwrapped.auth.client_token) do |client|
+          expect { client.logical.read("secret/test") }.to_not raise_error
+        end
+      end
+    end
+
+    describe "#unwrap_token", vault: ">= 0.6" do
+      it "returns the wrapped token when given a string" do
+        wrapped = vault_test_client.auth_token.create(wrap_ttl: "5s")
+        unwrapped = subject.unwrap_token(wrapped.wrap_info.token)
+
+        expect(unwrapped).to be
+
+        vault_test_client.with_token(unwrapped) do |client|
+          expect { client.logical.read("secret/test") }.to_not raise_error
+        end
+      end
+
+      it "returns the wrapped token when given a Vault::Secret" do
+        wrapped = vault_test_client.auth_token.create(wrap_ttl: "5s")
+        unwrapped = subject.unwrap_token(wrapped)
+
+        expect(unwrapped).to be
+
+        vault_test_client.with_token(unwrapped) do |client|
+          expect { client.logical.read("secret/test") }.to_not raise_error
+        end
+      end
+
+      it "returns nil when the response is empty" do
+        token = vault_test_client.auth_token.create # Note no wrap-ttl here
+        unwrapped = subject.unwrap_token(token.auth.client_token)
+        expect(unwrapped).to be(nil)
       end
     end
   end